moved Default dataclass from construction.py to parameters.py because it now has thermal environmental information and it is weird to have it under construction.py
